requirements.txt
pip~=24.0
attrs~=23.2.0
setuptools~=70.3.0
typing_extensions~=4.9.0
Mako~=1.3.2
MarkupSafe~=2.1.5
pytz~=2024.1
idna~=3.6
multidict~=6.0.4
aiogram~=3.3.0
pydantic~=2.5.3
aiofiles~=23.2.1
certifi~=2023.11.17
aiohttp~=3.9.1
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
SQLAlchemy~=2.0.25
alembic~=1.13.1
passlib~=1.7.4
urllib3~=2.2.1
bitarray~=2.9.2
pydantic_core~=2.14.6
pycparser~=2.21
cffi~=1.16.0
pytonconnect~=0.3.0
requests~=2.31.0
x25519~=0.0.2
python-dotenv~=1.0.1
pytoniq-core~=0.1.36
psycopg2-binary~=2.9.9
